fix(parser) : update parser to recognize endwised node
It is very difficult to recognize whether this node has end node.
Because the treesitter result when a endword is not existed is different
depends on languages/treesitter. It is annoying.

get_node() doesn't always give right node due to treesitter's wrapped
layer, So It is more convenient to detect root node of the region which
cursor in on. After then try to detect sub-treesitter.

In in_node(), it needs to get root node which has endwise, and find node
which is matched cursor position and target table.
To do this, check start_row of node is matched with cursor.
This code is not perfect and It will be modified as I said.

In is_endwised(), it needs to check root node which has endwised.
and check this rule to recognize endwised already.
1) check node has endword
2) check all nodes under root node have independent end row.

You must set proper node_name in configuration which can have end node.
It needs to execute node:end_().

It you use treesitter which use 'ERROR' to named_node like matlab, you
need to add 'ERROR' node in node_name table when you configure.