fix(parser) : bug fix for one line endwised statement
1. Problem

endwise_cr() detects that this block is not endwised by some rules.
One of them is checking whether endword is existed at the end of block.
Previously, the pattern checks when white space is inserted in front of
endword only.

It makes incorrect judgment that there are un-endwised statement
in the block when statements in block are has endword on one line.
For example, these sentences is recognized that there are no endword
`if a==b then c=1 end`
`function() a=1 end``

but I had to set this pattern like ('^%s+' .. endword) to deal with
situation that endword is commented like this.
if a==b then
    a=1
-- end
If  I set the pattern like ('%s+' .. endword), plugin will recognize
this block is correctly endwised.

2. Solution

Using commentstring, We add a step to detect endwise.
In case of block comment, endword will be in front of block
commentstring. so we needs to consider line-wise comment only.

1) using '%s*', check there are endword at end of block.
2) It yes, using commentstring, check this endword is commented.

This way is more reliable than before.